+++ src/Document.js
@@ -218,12 +218,12 @@ define(function (require, exports, module) {
 
     function init() {
         /*jshint validthis: true */
-        Channel.Document.comply(Events.CMD_HIGHLIGHT_CITES, _.bind(_handleHighlighting, this));
-        Channel.Extension.on(Events.EVT_PANEL_SHOWN,        _.bind(_handleFileScanning, this));
-        Channel.Extension.on(Events.EVT_EDITOR_CHANGE,      _.bind(_handleFileScanning, this));
+        Channel.Document.comply(Events.CMD_HIGHLIGHT_CITES,   _.bind(_handleHighlighting, this));
+        Channel.Extension.on(Events.EVT_PANEL_SHOWN,          _.bind(_handleFileScanning, this));
+        Channel.Extension.on(Events.EVT_EDITOR_CHANGE,        _.bind(_handleFileScanning, this));
 
-        Channel.Extension.comply(C.CMD_ID_INSERT_CITE,      _handleCiteInsertion);
-        Channel.Extension.comply(C.CMD_ID_INSERT_BIBLIO,    _handleBiblioInsertion);
+        Channel.Extension.comply(Events.CMD_INSERT_CITE,      _handleCiteInsertion);
+        Channel.Extension.comply(Events.CMD_INSERT_BIBLIO,    _handleBiblioInsertion);
     }
 
     function Document() {